ollama_back/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-cors==3.0.10
flask-pymongo==2.3.0
torch==1.13.1
requests==2.28.1
langchain==0.1.10
openai
pypdf==3.8.1
qdrant-client
tiktoken
crewai
pydantic==1.10.21
langchain-community
langchain-core
langchain-openai
python-dotenv
langchain_qdrant
fastapi